pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fysearch"
version = "0.1.0"
description = "CPU-only offline multimodal forensic search engine (OCR + embeddings + vector search)."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Rachit Pandey"}]

dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "numpy>=1.24",
  "pillow>=10.0",
  "pypdf>=4.0",
]

[project.optional-dependencies]
# Fast ANN search (recommended). If unavailable, code falls back to brute-force numpy.
faiss = ["faiss-cpu>=1.7.4"]

# Embeddings backends.
embeddings = [
  "sentence-transformers>=2.6",
  "torch>=2.1",
]

# OCR (requires system tesseract binary installed separately)
ocr = ["pytesseract>=0.3.10"]

# PDF rasterization for scanned PDFs (requires poppler installed separately)
pdf_images = ["pdf2image>=1.17"]

# Local web UI
web = ["flask>=3.0"]

[project.scripts]
fysearch = "fysearch.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
